+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# 1.5.0 (to be released)
+
+### Added
+
+- Support for multiple configurations and multiple repositories. See [cli documentation](docs/cli.md) for details
+- Possibility to use preconfigured Saxon `Processor` instance for validation
+
+### Changed
+
+- [ResolvingConfigurationStrategy.java#getProcessor()](de/kosit/validationtool/api/ResolvingConfigurationStrategy) is removed.
+- [INTERNAL] CLI parsing based on pico-cli, commons-cli is removed
+
 ## 1.4.2
 
 ### Fixed